Verb Function Module
Moved the verb functions (say, examine, and hit) from the main file into
verbFuncMod.py
Functions are accessed through VFM
